Start cleaning things up (#1)
