+++ covid_bot/utils/codes.py
@@ -1,0 +1,604 @@
+US_STATE_CODES = {
+    'AK': 'Alaska',
+    'AL': 'Alabama',
+    'AR': 'Arkansas',
+    'AS': 'American Samoa',
+    'AZ': 'Arizona',
+    'CA': 'California',
+    'CO': 'Colorado',
+    'CT': 'Connecticut',
+    'DC': 'District Of Columbia',
+    'DE': 'Delaware',
+    'FL': 'Florida',
+    'GA': 'Georgia',
+    'GU': 'Guam',
+    'HI': 'Hawaii',
+    'IA': 'Iowa',
+    'ID': 'Idaho',
+    'IL': 'Illinois',
+    'IN': 'Indiana',
+    'KS': 'Kansas',
+    'KY': 'Kentucky',
+    'LA': 'Louisiana',
+    'MA': 'Massachusetts',
+    'MD': 'Maryland',
+    'ME': 'Maine',
+    'MI': 'Michigan',
+    'MN': 'Minnesota',
+    'MO': 'Missouri',
+    'MP': 'Northern Mariana Islands',
+    'MS': 'Mississippi',
+    'MT': 'Montana',
+    'NA': 'National',
+    'NC': 'North Carolina',
+    'ND': 'North Dakota',
+    'NE': 'Nebraska',
+    'NH': 'New Hampshire',
+    'NJ': 'New Jersey',
+    'NM': 'New Mexico',
+    'NV': 'Nevada',
+    'NY': 'New York',
+    'OH': 'Ohio',
+    'OK': 'Oklahoma',
+    'OR': 'Oregon',
+    'PA': 'Pennsylvania',
+    'PR': 'Puerto Rico',
+    'RI': 'Rhode Island',
+    'SC': 'South Carolina',
+    'SD': 'South Dakota',
+    'TN': 'Tennessee',
+    'TX': 'Texas',
+    'UT': 'Utah',
+    'VA': 'Virginia',
+    'VI': 'Virgin Islands',
+    'VT': 'Vermont',
+    'WA': 'Washington',
+    'WI': 'Wisconsin',
+    'WV': 'West Virginia',
+    'WY': 'Wyoming',
+}
+
+# Mapping of country names to ISO 3166-1 alpha-2 codes
+COUNTRY_CODES_2 = {
+    'AF': 'Afghanistan',
+    'AX': 'Aland Islands',
+    'AL': 'Albania',
+    'DZ': 'Algeria',
+    'AS': 'American Samoa',
+    'AD': 'Andorra',
+    'AO': 'Angola',
+    'AI': 'Anguilla',
+    'AQ': 'Antarctica',
+    'AG': 'Antigua and Barbuda',
+    'AR': 'Argentina',
+    'AM': 'Armenia',
+    'AW': 'Aruba',
+    'AU': 'Australia',
+    'AT': 'Austria',
+    'AZ': 'Azerbaijan',
+    'BS': 'Bahamas',
+    'BH': 'Bahrain',
+    'BD': 'Bangladesh',
+    'BB': 'Barbados',
+    'BY': 'Belarus',
+    'BE': 'Belgium',
+    'BZ': 'Belize',
+    'BJ': 'Benin',
+    'BM': 'Bermuda',
+    'BT': 'Bhutan',
+    'BO': 'Bolivia',
+    'BQ': 'Bonaire',
+    'BA': 'Bosnia and Herzegovina',
+    'BW': 'Botswana',
+    'BV': 'Bouvet Island',
+    'BR': 'Brazil',
+    'IO': 'British Indian Ocean Territory',
+    'BN': 'Brunei',
+    'BG': 'Bulgaria',
+    'BF': 'Burkina Faso',
+    'BI': 'Burundi',
+    'KH': 'Cambodia',
+    'CM': 'Cameroon',
+    'CA': 'Canada',
+    'CV': 'Cape Verde',
+    'KY': 'Cayman Islands',
+    'CF': 'Central African Republic',
+    'TD': 'Chad',
+    'CL': 'Chile',
+    'CN': 'China',
+    'CX': 'Christmas Island',
+    'CC': 'Cocos (Keeling) Islands',
+    'CO': 'Colombia',
+    'KM': 'Comoros',
+    'CG': 'Congo',
+    'CD': 'Congo, the Democratic Republic of the',
+    'CK': 'Cook Islands',
+    'CR': 'Costa Rica',
+    'CI': 'Cote d\'Ivoire',
+    'HR': 'Croatia',
+    'CU': 'Cuba',
+    'CW': 'Curaçao',
+    'CY': 'Cyprus',
+    'CZ': 'Czechia',
+    'DK': 'Denmark',
+    'DJ': 'Djibouti',
+    'DM': 'Dominica',
+    'DO': 'Dominican Republic',
+    'EC': 'Ecuador',
+    'EG': 'Egypt',
+    'SV': 'El Salvador',
+    'GQ': 'Equatorial Guinea',
+    'ER': 'Eritrea',
+    'EE': 'Estonia',
+    'ET': 'Ethiopia',
+    'FK': 'Falkland Islands (Malvinas)',
+    'FO': 'Faroe Islands',
+    'FJ': 'Fiji',
+    'FI': 'Finland',
+    'FR': 'France',
+    'GF': 'French Guiana',
+    'PF': 'French Polynesia',
+    'TF': 'French Southern Territories',
+    'GA': 'Gabon',
+    'GM': 'Gambia',
+    'GE': 'Georgia',
+    'DE': 'Germany',
+    'GH': 'Ghana',
+    'GI': 'Gibraltar',
+    'GR': 'Greece',
+    'GL': 'Greenland',
+    'GD': 'Grenada',
+    'GP': 'Guadeloupe',
+    'GU': 'Guam',
+    'GT': 'Guatemala',
+    'GG': 'Guernsey',
+    'GN': 'Guinea',
+    'GW': 'Guinea-Bissau',
+    'GY': 'Guyana',
+    'HT': 'Haiti',
+    'HM': 'Heard Island and McDonald Islands',
+    'VA': 'Holy See',
+    'HN': 'Honduras',
+    'HK': 'Hong Kong',
+    'HU': 'Hungary',
+    'IS': 'Iceland',
+    'IN': 'India',
+    'ID': 'Indonesia',
+    'IR': 'Iran',
+    'IQ': 'Iraq',
+    'IE': 'Ireland',
+    'IM': 'Isle of Man',
+    'IL': 'Israel',
+    'IT': 'Italy',
+    'JM': 'Jamaica',
+    'JP': 'Japan',
+    'JE': 'Jersey',
+    'JO': 'Jordan',
+    'KZ': 'Kazakhstan',
+    'KE': 'Kenya',
+    'KI': 'Kiribati',
+    'KP': 'Democratic People\'s Republic of Korea',
+    'KR': 'S. Korea',
+    'KW': 'Kuwait',
+    'KG': 'Kyrgyzstan',
+    'LA': 'Lao People\'s Democratic Republic',
+    'LV': 'Latvia',
+    'LB': 'Lebanon',
+    'LS': 'Lesotho',
+    'LR': 'Liberia',
+    'LY': 'Libya',
+    'LI': 'Liechtenstein',
+    'LT': 'Lithuania',
+    'LU': 'Luxembourg',
+    'MO': 'Macau',
+    'MK': 'North Macedonia',
+    'MG': 'Madagascar',
+    'MW': 'Malawi',
+    'MY': 'Malaysia',
+    'MV': 'Maldives',
+    'ML': 'Mali',
+    'MT': 'Malta',
+    'MH': 'Marshall Islands',
+    'MQ': 'Martinique',
+    'MR': 'Mauritania',
+    'MU': 'Mauritius',
+    'YT': 'Mayotte',
+    'MX': 'Mexico',
+    'FM': 'Federated States of Micronesia',
+    'MD': 'Moldova',
+    'MC': 'Monaco',
+    'MN': 'Mongolia',
+    'ME': 'Montenegro',
+    'MS': 'Montserrat',
+    'MA': 'Morocco',
+    'MZ': 'Mozambique',
+    'MM': 'Myanmar',
+    'NA': 'Namibia',
+    'NR': 'Nauru',
+    'NP': 'Nepal',
+    'NL': 'Netherlands',
+    'NC': 'New Caledonia',
+    'NZ': 'New Zealand',
+    'NI': 'Nicaragua',
+    'NE': 'Niger',
+    'NG': 'Nigeria',
+    'NU': 'Niue',
+    'NF': 'Norfolk Island',
+    'MP': 'Northern Mariana Islands',
+    'NO': 'Norway',
+    'OM': 'Oman',
+    'PK': 'Pakistan',
+    'PW': 'Palau',
+    'PS': 'Palestine',
+    'PA': 'Panama',
+    'PG': 'Papua New Guinea',
+    'PY': 'Paraguay',
+    'PE': 'Peru',
+    'PH': 'Philippines',
+    'PN': 'Pitcairn',
+    'PL': 'Poland',
+    'PT': 'Portugal',
+    'PR': 'Puerto Rico',
+    'QA': 'Qatar',
+    'RE': 'Reunion',
+    'RO': 'Romania',
+    'RU': 'Russia',
+    'RW': 'Rwanda',
+    'BL': 'St. Barth',
+    'SH': 'Saint Helena, Ascension and Tristan da Cunha',
+    'KN': 'Saint Kitts and Nevis',
+    'LC': 'Saint Lucia',
+    'MF': 'Saint Martin',
+    'PM': 'Saint Pierre and Miquelon',
+    'VC': 'Saint Vincent and the Grenadines',
+    'WS': 'Samoa',
+    'SM': 'San Marino',
+    'ST': 'Sao Tome and Principe',
+    'SA': 'Saudi Arabia',
+    'SN': 'Senegal',
+    'RS': 'Serbia',
+    'SC': 'Seychelles',
+    'SL': 'Sierra Leone',
+    'SG': 'Singapore',
+    'SX': 'Sint Maarten',
+    'SK': 'Slovakia',
+    'SI': 'Slovenia',
+    'SB': 'Solomon Islands',
+    'SO': 'Somalia',
+    'ZA': 'South Africa',
+    'GS': 'South Georgia and the South Sandwich Islands',
+    'SS': 'South Sudan',
+    'ES': 'Spain',
+    'LK': 'Sri Lanka',
+    'SD': 'Sudan',
+    'SR': 'Suriname',
+    'SJ': 'Svalbard and Jan Mayen',
+    'SZ': 'Swaziland',
+    'SE': 'Sweden',
+    'CH': 'Switzerland',
+    'SY': 'Syria',
+    'TW': 'Taiwan',
+    'TJ': 'Tajikistan',
+    'TZ': 'Tanzania',
+    'TH': 'Thailand',
+    'TL': 'Timor-Leste',
+    'TG': 'Togo',
+    'TK': 'Tokelau',
+    'TO': 'Tonga',
+    'TT': 'Trinidad and Tobago',
+    'TN': 'Tunisia',
+    'TR': 'Turkey',
+    'TM': 'Turkmenistan',
+    'TC': 'Turks and Caicos Islands',
+    'TV': 'Tuvalu',
+    'UG': 'Uganda',
+    'UA': 'Ukraine',
+    'AE': 'United Arab Emirates',
+    'GB': 'UK',
+    'US': 'USA',
+    'UM': 'United States Minor Outlying Islands',
+    'UY': 'Uruguay',
+    'UZ': 'Uzbekistan',
+    'VU': 'Vanuatu',
+    'VE': 'Venezuela',
+    'VN': 'Vietnam',
+    'VG': 'British Virgin Islands',
+    'VI': 'Virgin Islands, U.S.',
+    'WF': 'Wallis and Futuna',
+    'EH': 'Western Sahara',
+    'YE': 'Yemen',
+    'ZM': 'Zambia',
+    'ZW': 'Zimbabwe',
+}
+
+# Mapping of country names to ISO 3166-1 alpha-2 codes
+COUNTRY_CODES_3 = {
+    'ABW': 'Aruba',
+    'AFG': 'Afghanistan',
+    'AGO': 'Angola',
+    'AIA': 'Anguilla',
+    'ALA': 'Åland Islands',
+    'ALB': 'Albania',
+    'AND': 'Andorra',
+    'ARE': 'United Arab Emirates',
+    'ARG': 'Argentina',
+    'ARM': 'Armenia',
+    'ASM': 'American Samoa',
+    'ATA': 'Antarctica',
+    'ATF': 'French Southern Territories',
+    'ATG': 'Antigua and Barbuda',
+    'AUS': 'Australia',
+    'AUT': 'Austria',
+    'AZE': 'Azerbaijan',
+    'BDI': 'Burundi',
+    'BEL': 'Belgium',
+    'BEN': 'Benin',
+    'BES': 'Bonaire, Sint Eustatius and Saba',
+    'BFA': 'Burkina Faso',
+    'BGD': 'Bangladesh',
+    'BGR': 'Bulgaria',
+    'BHR': 'Bahrain',
+    'BHS': 'Bahamas',
+    'BIH': 'Bosnia and Herzegovina',
+    'BLM': 'St. Barth',
+    'BLR': 'Belarus',
+    'BLZ': 'Belize',
+    'BMU': 'Bermuda',
+    'BOL': 'Bolivia',
+    'BRA': 'Brazil',
+    'BRB': 'Barbados',
+    'BRN': 'Brunei',
+    'BTN': 'Bhutan',
+    'BVT': 'Bouvet Island',
+    'BWA': 'Botswana',
+    'CAF': 'Central African Republic',
+    'CAN': 'Canada',
+    'CCK': 'Cocos (Keeling) Islands',
+    'CHE': 'Switzerland',
+    'CHL': 'Chile',
+    'CHN': 'China',
+    'CIV': 'Cote d\'Ivoire',
+    'CMR': 'Cameroon',
+    'COD': 'Congo (the Democratic Republic of the)',
+    'COG': 'Congo',
+    'COK': 'Cook Islands',
+    'COL': 'Colombia',
+    'COM': 'Comoros',
+    'CPV': 'Cabo Verde',
+    'CRI': 'Costa Rica',
+    'CUB': 'Cuba',
+    'CUW': 'Curaçao',
+    'CXR': 'Christmas Island',
+    'CYM': 'Cayman Islands',
+    'CYP': 'Cyprus',
+    'CZE': 'Czechia',
+    'DEU': 'Germany',
+    'DJI': 'Djibouti',
+    'DMA': 'Dominica',
+    'DNK': 'Denmark',
+    'DOM': 'Dominican Republic',
+    'DZA': 'Algeria',
+    'ECU': 'Ecuador',
+    'EGY': 'Egypt',
+    'ERI': 'Eritrea',
+    'ESH': 'Western Sahara*',
+    'ESP': 'Spain',
+    'EST': 'Estonia',
+    'ETH': 'Ethiopia',
+    'FIN': 'Finland',
+    'FJI': 'Fiji',
+    'FLK': 'Falkland Islands [Malvinas]',
+    'FRA': 'France',
+    'FRO': 'Faroe Islands',
+    'FSM': 'Micronesia (Federated States of)',
+    'GAB': 'Gabon',
+    'GBR': 'UK',
+    'GEO': 'Georgia',
+    'GGY': 'Guernsey',
+    'GHA': 'Ghana',
+    'GIB': 'Gibraltar',
+    'GIN': 'Guinea',
+    'GLP': 'Guadeloupe',
+    'GMB': 'Gambia',
+    'GNB': 'Guinea-Bissau',
+    'GNQ': 'Equatorial Guinea',
+    'GRC': 'Greece',
+    'GRD': 'Grenada',
+    'GRL': 'Greenland',
+    'GTM': 'Guatemala',
+    'GUF': 'French Guiana',
+    'GUM': 'Guam',
+    'GUY': 'Guyana',
+    'HKG': 'Hong Kong',
+    'HMD': 'Heard Island and McDonald Islands',
+    'HND': 'Honduras',
+    'HRV': 'Croatia',
+    'HTI': 'Haiti',
+    'HUN': 'Hungary',
+    'IDN': 'Indonesia',
+    'IMN': 'Isle of Man',
+    'IND': 'India',
+    'IOT': 'British Indian Ocean Territory',
+    'IRL': 'Ireland',
+    'IRN': 'Iran',
+    'IRQ': 'Iraq',
+    'ISL': 'Iceland',
+    'ISR': 'Israel',
+    'ITA': 'Italy',
+    'JAM': 'Jamaica',
+    'JEY': 'Jersey',
+    'JOR': 'Jordan',
+    'JPN': 'Japan',
+    'KAZ': 'Kazakhstan',
+    'KEN': 'Kenya',
+    'KGZ': 'Kyrgyzstan',
+    'KHM': 'Cambodia',
+    'KIR': 'Kiribati',
+    'KNA': 'Saint Kitts and Nevis',
+    'KOR': 'S. Korea',
+    'KWT': 'Kuwait',
+    'LAO': 'Lao People\'s Democratic Republic',
+    'LBN': 'Lebanon',
+    'LBR': 'Liberia',
+    'LBY': 'Libya',
+    'LCA': 'Saint Lucia',
+    'LIE': 'Liechtenstein',
+    'LKA': 'Sri Lanka',
+    'LSO': 'Lesotho',
+    'LTU': 'Lithuania',
+    'LUX': 'Luxembourg',
+    'LVA': 'Latvia',
+    'MAC': 'Macau',
+    'MAF': 'Saint Martin',
+    'MAR': 'Morocco',
+    'MCO': 'Monaco',
+    'MDA': 'Moldova',
+    'MDG': 'Madagascar',
+    'MDV': 'Maldives',
+    'MEX': 'Mexico',
+    'MHL': 'Marshall Islands',
+    'MKD': 'North Macedonia',
+    'MLI': 'Mali',
+    'MLT': 'Malta',
+    'MMR': 'Myanmar',
+    'MNE': 'Montenegro',
+    'MNG': 'Mongolia',
+    'MNP': 'Northern Mariana Islands',
+    'MOZ': 'Mozambique',
+    'MRT': 'Mauritania',
+    'MSR': 'Montserrat',
+    'MTQ': 'Martinique',
+    'MUS': 'Mauritius',
+    'MWI': 'Malawi',
+    'MYS': 'Malaysia',
+    'MYT': 'Mayotte',
+    'NAM': 'Namibia',
+    'NCL': 'New Caledonia',
+    'NER': 'Niger',
+    'NFK': 'Norfolk Island',
+    'NGA': 'Nigeria',
+    'NIC': 'Nicaragua',
+    'NIU': 'Niue',
+    'NLD': 'Netherlands',
+    'NOR': 'Norway',
+    'NPL': 'Nepal',
+    'NRU': 'Nauru',
+    'NZL': 'New Zealand',
+    'OMN': 'Oman',
+    'PAK': 'Pakistan',
+    'PAN': 'Panama',
+    'PCN': 'Pitcairn',
+    'PER': 'Peru',
+    'PHL': 'Philippines',
+    'PLW': 'Palau',
+    'PNG': 'Papua New Guinea',
+    'POL': 'Poland',
+    'PRI': 'Puerto Rico',
+    'PRK': 'Korea (the Democratic People\'s Republic of)',
+    'PRT': 'Portugal',
+    'PRY': 'Paraguay',
+    'PSE': 'Palestine',
+    'PYF': 'French Polynesia',
+    'QAT': 'Qatar',
+    'REU': 'Reunion',
+    'ROU': 'Romania',
+    'RUS': 'Russia',
+    'RWA': 'Rwanda',
+    'SAU': 'Saudi Arabia',
+    'SDN': 'Sudan',
+    'SEN': 'Senegal',
+    'SGP': 'Singapore',
+    'SGS': 'South Georgia and the South Sandwich Islands',
+    'SHN': 'Saint Helena, Ascension and Tristan da Cunha',
+    'SJM': 'Svalbard and Jan Mayen',
+    'SLB': 'Solomon Islands',
+    'SLE': 'Sierra Leone',
+    'SLV': 'El Salvador',
+    'SMR': 'San Marino',
+    'SOM': 'Somalia',
+    'SPM': 'Saint Pierre and Miquelon',
+    'SRB': 'Serbia',
+    'SSD': 'South Sudan',
+    'STP': 'Sao Tome and Principe',
+    'SUR': 'Suriname',
+    'SVK': 'Slovakia',
+    'SVN': 'Slovenia',
+    'SWE': 'Sweden',
+    'SWZ': 'Eswatini',
+    'SXM': 'Sint Maarten',
+    'SYC': 'Seychelles',
+    'SYR': 'Syria',
+    'TCA': 'Turks and Caicos Islands',
+    'TCD': 'Chad',
+    'TGO': 'Togo',
+    'THA': 'Thailand',
+    'TJK': 'Tajikistan',
+    'TKL': 'Tokelau',
+    'TKM': 'Turkmenistan',
+    'TLS': 'Timor-Leste',
+    'TON': 'Tonga',
+    'TTO': 'Trinidad and Tobago',
+    'TUN': 'Tunisia',
+    'TUR': 'Turkey',
+    'TUV': 'Tuvalu',
+    'TWN': 'Taiwan',
+    'TZA': 'Tanzania',
+    'UGA': 'Uganda',
+    'UKR': 'Ukraine',
+    'UMI': 'United States Minor Outlying Islands',
+    'URY': 'Uruguay',
+    'USA': 'USA',
+    'UZB': 'Uzbekistan',
+    'VAT': 'Holy See',
+    'VCT': 'Saint Vincent and the Grenadines',
+    'VEN': 'Venezuela',
+    'VGB': 'British Virgin Islands',
+    'VIR': 'Virgin Islands (U.S.)',
+    'VNM': 'Vietnam',
+    'VUT': 'Vanuatu',
+    'WLF': 'Wallis and Futuna',
+    'WSM': 'Samoa',
+    'YEM': 'Yemen',
+    'ZAF': 'South Africa',
+    'ZMB': 'Zambia',
+    'ZWE': 'Zimbabwe',
+}
+
+# Other names for countries
+ALT_NAMES = {
+    'Czech Republic': 'Czechia',
+    'United Kingdom': 'UK',
+    'South Korea': 'S. Korea',
+    'Korea': 'S. Korea',
+    'United States': 'USA',
+}
+
+JHU_NAMES = {
+    'USA': 'US',
+    'S. Korea': 'Korea, South',
+    'UK': 'United Kingdom',
+}
+
+# Alternate lookups
+US_STATE_CODES_VALUES = set(US_STATE_CODES.values())
+COUNTRY_CODES_2_VALUES = set(COUNTRY_CODES_2.values())
+COUNTRY_CODES_3_VALUES = set(COUNTRY_CODES_3.values())
+JHU_NAMES_VALUES = set(JHU_NAMES.values())
+
+# Magical Emoji codes used for embeds
+EMOJI_CODES = {
+    'discord': '<:discord:689486285349715995>',
+    'github': '<:github:689501322969350158>',
+    'kofi': '<:Kofi:689483361785217299>',
+    'server': '<:servers:689502498251341953>',
+    'upvote': '<:upvote:689186080070959207>',
+    'user': '<:user:689502620590669912>',
+
+    # extra
+    'confirmed': '<:confirmed:689494326493184090>',
+    'deaths': '<:deaths:689489690101153800>',
+    'recovered': '<:recovered:689490988808274003>',
+    'activecases': '<:activecases:689494177733410861>',
+    'mortalityrate': '<:mortalityrate:689488380865544345>',
+    'recoveryrate': '<:recoveryrate:689492820125417521>',
+}
