+++ covid_bot/utils/time.py
@@ -1,0 +1,7 @@
+import datetime
+
+
+def utcnow():
+    """ Return a non-naive UTC datetime object
+    """
+    return datetime.datetime.utcnow(datetime.timezone.utc)
