+++ setup.cfg
@@ -1,0 +1,24 @@
+# Documentation for the contents of this file can be found here:
+# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+
+[metadata]
+name = covid_bot
+version = 0.0.1
+description = Covid-19 Discord Bot
+url = https://github.com/picklejason/coronavirus-bot
+license = GPL 3.0
+classifiers =
+    Development Status :: 4 - Beta
+    Environment :: Console
+    Intended Audience :: Developers
+    License :: OSI Approved :: GPL 3.0
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Topic :: Software Development :: Libraries
+    Operating System :: POSIX
+    Operating System :: Unix
+    Operating System :: MacOS
+
+[options]
+zip_safe = False
+packages = find:
