+++ setup.py
@@ -1,0 +1,6 @@
+from setuptools import setup
+
+
+if __name__ == '__main__':
+    # See setup.cfg for package metadata
+    setup()
